[KOGITO-3783] Make more components optional in StaticServiceRegistry
